iete/requirements.txt
numpy
pandas
yfinance
streamlit
plotly
requests
statsmodels
matplotlib
phi-agent
phi-tools
duckduckgo-search
duckdb
